Get rid of StatsThread
Heavy refactoring here. HdbStat were moved away from PushThread. All the
information is now on the HdbSignal class.

all the statistics are updated as they go, note that a global lock on
the reading for AttributeRecordFreqList and the same with failure is
required but not set yet.
All signals can now be operated on separate threads for maximum
efficiency, but the stats need a global lock for the list of statistics.
As such it is not completely implemented, the events have been dropped,
but could be reintroduced as well.